Rewrote encode_bdict, destroy_bdict, removed bdict_stack
 Changes to be committed:
	modified:   bencode.c
	modified:   bencode.h

I rewrote encode_bdict and destroy_bdict because they were unnecessarily
complicated. This also eliminated the need for all the bdict_stack code,
so I eliminated that too. However, there is a problem with the bencoded
dictionary parser which is that it doesn't store the length of strings.
So the only way to get the length is by using strlen, but of course that
only works if the string is NULL-terminated. The one string in most
bencoded dictionaries that can't be NULL-terminated is the "pieces"
entry. Since it's a binary string (it's a concatenation of the hashes
of the various pieces the file is made up of), it can't be NULL-terminated,
since a NULL value could easily occur within it. Thus, the length of the
"pieces" string has to be stored somewhere. What I previously did, was find
the length of the file in the bencoded dictionary, and then just found the
length of the pieeces string with the formula: ceil(length, piece_length)*20.
However, the "length" entry is not present in every torrent file. Thus, I need
to modify struct bdict to store the length of the string (if it stores a string).

Since I'm already going to be modifying struct bdict, I'm also going to switch to
arbitrary-precision integers, since the Bittorrent specification says that integers,
and strings in bencoded dictionaries can be arbitrarily large. I previously used 64-bit
integers, which would probably be fine for most torrents, but I want this program
to be fully compliant with the Bittorrent Specification.