Fixed bencode parser to store length of strings
Prior to this commit, the bdict data structure didn't
store the length of strings. However, this didn't work
all the time since strings in bencoded dictionaries
aren't always null-terminated. For example, the "pieces"
string, which is present in every torrent metafile, isn't
null-terminated. So prior to this commit, encoding the pieces
string required special handling. The bencode encoder would look
for a "length" string in the bencoded dictionary in order to
get the length of the string. However, this didn't work if the
torrent metafile tracked a directory instead of a regular file. In
that case, each file to be downloaded would have it's own length
string. Therefore, it's much easier to just store the length of
every string.

I am planning on eventually switching from 64-bit integers to
arbitrarily large ones. I'll be using the fgmp library which is
a public-domain re-implementation of the gnu gmp library. It's
considerably slower than the GNU version, but I'm only using fgmp
since I want this program to conform to the BEP 3 specification.
Most of the integers in bencoded dictionaries will probably
be less than 64 bits anyway, so using fgmp probably won't affect
performance too much. If it does, I might switch to the GNU gmp
library.

 Changes to be committed:
	modified:   Makefile
	modified:   bencode.c
	modified:   bencode.h
	new file:   fgmp/Makefile
	new file:   fgmp/gmp.c
	new file:   fgmp/gmp.h
	new file:   fgmp/notes
	modified:   torrent.c